% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo-extras.R
\name{add_legend}
\alias{add_legend}
\title{Legend for topogram}
\usage{
add_legend(topo, title, n_cells = 10, label_format = ".2f",
  labels = NULL, orientation = "horizontal", cells_width = 30,
  cells_height = 15, cells_padding = 2, title_width = 100)
}
\arguments{
\item{topo}{A \code{topogram} \code{htmlwidget} object.}

\item{title}{Legend title}

\item{n_cells}{Number of cells in the legend.}

\item{label_format}{A string passed to \code{d3.format},
see \url{https://github.com/d3/d3-format}.}

\item{labels}{A character vector to use as labels, better if same length as number of cells.}

\item{orientation}{Legend orientation: vertical or horizontal, can be abbreviated.}

\item{cells_width}{Width of each cells.}

\item{cells_height}{Height of each cells.}

\item{cells_padding}{Space between each cells.}

\item{title_width}{Width for the title.}
}
\description{
Legend for topogram
}
\examples{
library(topogram)
data(paris)

topogram(
  shape = paris,
  value = "TOTAL",
  n_iteration = 10
) \%>\% add_legend(
  title = "Population",
  n_cells = 12,
  orientation = "vertical"
)


# Format labels favue with D3 format
topogram(
  shape = paris,
  value = "TOTAL",
  n_iteration = 10
) \%>\% add_legend(
  title = "Population",
  label_format = ".2s",
  orientation = "vertical"
)


# No space between cells
# and custom labels
topogram(
  shape = paris,
  value = "TOTAL",
  n_iteration = 10
) \%>\% add_legend(
  title = "Population",
  n_cells = 20,
  cells_padding = 0,
  cells_width = 10,
  cells_height = 20,
  labels = c("15 000", rep("", 18), "250 000"),
  orientation = "vertical"
)
}
