% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{france}
\alias{france}
\title{France departments}
\format{
An 'sf' object containing geometries and data.
}
\source{
Natural Earth (via package {rnaturalearth}) for polygons, INSEE for data.
}
\usage{
france
}
\description{
France departments with data about structure of total employment by major industry in 2019.
}
\keyword{datasets}
