% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paris}
\alias{paris}
\title{Population of Paris by age in 2014}
\format{
A data frame with 20 observations on the following 15 variables.
\describe{
  \item{\code{CODE_INSEE}}{Code of the district}
  \item{\code{LIB}}{Label of the district}
  \item{\code{NAME}}{Name of the district}
  \item{\code{AGE_00}}{Population under 3 years old}
  \item{\code{AGE_03}}{Population of people aged 3-5 years}
  \item{\code{AGE_06}}{Population of people aged 6-10 years}
  \item{\code{AGE_11}}{Population of people aged 11-17 years}
  \item{\code{AGE_18}}{Population of people aged 18-24 years}
  \item{\code{AGE_25}}{Population of people aged 25-39 years}
  \item{\code{AGE_40}}{Population of people aged 40-54 years}
  \item{\code{AGE_55}}{Population of people aged 55-64 years}
  \item{\code{AGE_65}}{Population of people aged 65-79 years}
  \item{\code{AGE_80}}{Population of people aged 80 and over}
  \item{\code{TOTAL}}{Population total}
  \item{\code{geometry}}{a sfc_POLYGON}
}
}
\source{
INSEE (\url{https://www.insee.fr/fr/statistiques/1893198}) and \url{opendata.paris.fr}
}
\usage{
paris
}
\description{
An 'sf' object containing geometries and data.
}
\keyword{datasets}
