% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_france}
\alias{pop_france}
\title{French population estimate as at 1er January 2018}
\format{A data frame with 101 observations on the following 20 variables.
\describe{
  \item{\code{departements_code}}{Department code}
  \item{\code{departements_lib}}{Department label}
  \item{\code{ensemble_0_a_19_ans}}{Total population of people aged 0-19 years}
  \item{\code{ensemble_20_a_39_ans}}{Total population of people aged 20-39 years}
  \item{\code{ensemble_40_a_59_ans}}{Total population of people aged 40-59 years}
  \item{\code{ensemble_60_a_74_ans}}{Total population of people aged 60-74 years}
  \item{\code{ensemble_75_ans_et_plus}}{Total population of people aged 75 and over}
  \item{\code{ensemble_total}}{Total population}
  \item{\code{hommes_0_a_19_ans}}{Total population of men aged 0-19 years}
  \item{\code{hommes_20_a_39_ans}}{Total population of men aged 20-39 years}
  \item{\code{hommes_40_a_59_ans}}{Total population of men aged 40-59 years}
  \item{\code{hommes_60_a_74_ans}}{Total population of men aged 60-74 years}
  \item{\code{hommes_75_ans_et_plus}}{Total population of men aged 75 and over}
  \item{\code{hommes_total}}{Total population of men}
  \item{\code{femmes_0_a_19_ans}}{Total population of women aged 0-19 years}
  \item{\code{femmes_20_a_39_ans}}{Total population of women aged 20-39 years}
  \item{\code{femmes_40_a_59_ans}}{Total population of women aged 40-59 years}
  \item{\code{femmes_60_a_74_ans}}{Total population of women aged 60-74 years}
  \item{\code{femmes_75_ans_et_plus}}{Total population of women aged 75 and over}
  \item{\code{femmes_total}}{Total population of women}
}}
\source{
INSEE (\url{https://www.insee.fr/fr/statistiques/1893198})
}
\usage{
pop_france
}
\description{
Data by department, gender and age.
}
\keyword{datasets}
