% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{proxy_update_iteration}
\alias{proxy_update_iteration}
\title{Update number of iteration to distort topology}
\usage{
proxy_update_iteration(proxy, n_iteration)
}
\arguments{
\item{proxy}{A \code{topogramProxy} \code{htmlwidget} object.}

\item{n_iteration}{Number of iteration to perform.}
}
\value{
A \code{topogramProxy} \code{htmlwidget} object.
}
\description{
Use this in 'shiny' application to update an already generated cartogram.
}
\examples{
if (interactive()) {



}
}
