% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topogram-package.R
\name{topogram-exports}
\alias{topogram-exports}
\alias{\%>\%}
\title{topogram exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the topogram package to avoid listing the magrittr
as Depends of topogram
}
