% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topogram-package.R
\docType{package}
\name{topogram-package}
\alias{topogram-package}
\title{An \code{htmlwidget} to create Continuous Cartogram in 'D3.js'}
\description{
This package allow you to use cartogram-chart.js (\url{https://github.com/vasturiano/cartogram-chart}),
an interactive contiguous cartogram reusable chart for visualizing geographical data.
}
\author{
Victor Perrier (@dreamRs_fr)
}
