% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{topogRam-shiny}
\alias{topogRam-shiny}
\alias{topogRamOutput}
\alias{renderTopogRam}
\alias{topogramProxy}
\title{Shiny bindings for topogRam}
\usage{
topogRamOutput(outputId, width = "100\%", height = "400px")

renderTopogRam(expr, env = parent.frame(), quoted = FALSE)

topogramProxy(shinyId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a topogRam}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{shinyId}{single-element character vector indicating the output ID of the
chart to modify (if invoked from a Shiny module, the namespace will be added
automatically)}

\item{session}{the Shiny session object to which the chart belongs; usually the
default value will suffice}
}
\description{
Output and render functions for using topogRam within Shiny
applications and interactive Rmd documents.
}
