% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{topogram_proxy_update}
\alias{topogram_proxy_update}
\title{Update variable used to distort topology with proxy}
\usage{
topogram_proxy_update(
  proxy,
  sfobj,
  value,
  label = "{value}",
  palette = "viridis"
)
}
\arguments{
\item{proxy}{A \code{topogram_proxy} \code{htmlwidget} object.}

\item{sfobj}{An \code{sf} object. For the time being, shape must be projected in Mercator (CRS 4326).}

\item{value}{Variable name to use to distort topology.}

\item{label}{\code{glue} string to be used in tooltip, you can use HTML tags in it.}

\item{palette}{Name of a color palette, such as \code{"viridis"}, \code{"Blues"}, ... Or a function to map data values to colors, see \code{\link[scales:col_numeric]{scales::col_numeric()}}.}
}
\value{
A \code{topogram_proxy} \code{htmlwidget} object.
}
\description{
Use this in 'shiny' application to update an already generated cartogram.
}
