% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topogram_select.R
\name{topogram_select}
\alias{topogram_select}
\title{Select menu to update \code{\link[=topogram]{topogram()}}}
\usage{
topogram_select(
  topogramId,
  sfobj,
  values,
  label = "{value}",
  palette = "viridis",
  rescale_to = c(1, 1000),
  n_iteration = 10,
  width = NULL
)
}
\arguments{
\item{topogramId}{The \code{elementId} of the \code{\link[=topogram]{topogram()}} to update.}

\item{sfobj}{An \code{sf} object. For the time being, shape must be projected in Mercator (CRS 4326).}

\item{values}{Parameters to construct cartograms, can be:
\itemize{
\item a \code{character} vector of variable to use
\item a named \code{list} where names will be used in select menu and values as variable
\item a \code{list} of \code{lists} where each sub-list can contain: \code{value} (variable), \code{text} (label for select menu),
\code{palette}, \code{labels} (parameters specific for the variable considered)
}}

\item{label}{\code{glue} string to be used in tooltip, you can use HTML tags in it.}

\item{palette}{Name of a color palette, such as \code{"viridis"}, \code{"Blues"}, ...
Or a function to map data values to colors, see \code{\link[scales:col_numeric]{scales::col_numeric()}}.}

\item{rescale_to}{Rescale value to distort topology to a specified range, use \code{NULL} to use values as is.}

\item{n_iteration}{Number of iterations to run the algorithm for. Higher numbers distorts the areas closer to their associated value,
at the cost of performance.}

\item{width}{A numeric input in pixels.}
}
\description{
Use in RMarkdown documents to update a \code{\link[=topogram]{topogram()}} dynamically.
}
\examples{
library(topogram)
library(htmltools)

# normally, you would use the following in RMarkdown

browsable(tagList(
  # Select menu
  topogram_select(
    topogramId = "ID",
    sfobj = world,
    values = list(
      "Population" = "pop_est",
      "GDP" = "gdp_md_est", 
      "CO2 emissions (1990)" = "co2_emissions_1990", 
      "CO2 emissions (2020)" = "co2_emissions_2020", 
      "Share of electricity production from renewables" = "renewables_percent_electricity"
    ),
    label = "{name} : {value}"
  ),
  
  # Topogram
  topogram(
    sfobj = world,
    value = "pop_est", 
    label = "{name} : {value}",
    elementId = "ID"
  )
), value = interactive())


# specific options according to variables
browsable(tagList(
  # Select menu
  topogram_select(
    topogramId = "ID",
    sfobj = world,
    values = list(
      list(
        text = "Population", 
        value = "pop_est", 
        palette = "Greens", 
        label = "{name} : {value} inhabitants"
      ),
      list(
        text = "GDP",
        value = "gdp_md_est", 
        palette = "Blues", 
        label = "{name} : ${value}"
      )
    ),
    label = "{name} : {value}"
  ),
  
  # Topogram
  topogram(
    sfobj = world,
    value = "pop_est", 
    elementId = "ID",
    # you have to specify parameters for initializing topogram
    palette = "Greens",
    label = "{name} : {value} inhabitants"
  )
), value = interactive())
}
