% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world}
\alias{world}
\title{World Countries}
\format{
An 'sf' object containing geometries and data.
}
\source{
Natural Earth (via package {rnaturalearth}) for polygons, Our World In Data for CO2 and renewables electricity data.
}
\usage{
world
}
\description{
World countries with data such as population estimate, GDP,
CO2 emissions and share of electricity production from renewables.
}
\keyword{datasets}
