% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo-extras.R
\name{add_labs}
\alias{add_labs}
\title{Labs for topogRam}
\usage{
add_labs(topo, title = NULL, subtitle = NULL, caption = NULL)
}
\arguments{
\item{topo}{A \code{topogRam} \code{htmlwidget} object.}

\item{title}{Main title.}

\item{subtitle}{Subtitle.}

\item{caption}{Brief explanation of the source of the data.}
}
\description{
Labs for topogRam
}
\examples{
#TODO
}
