% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{proxy_update_value}
\alias{proxy_update_value}
\title{Update variable used to distort topology with proxy}
\usage{
proxy_update_value(proxy, new_value)
}
\arguments{
\item{proxy}{A \code{topogramProxy} \code{htmlwidget} object.}

\item{new_value}{New variable to use, must a \code{character} of length one,
and a valid variable name of data used to construct the cartogram.}
}
\value{
A \code{topogramProxy} \code{htmlwidget} object.
}
\description{
Use this in 'shiny' application to update an already generated cartogram.
}
\examples{
if (interactive()) {



}
}
