% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topogRam-package.R
\name{topogRam-exports}
\alias{topogRam-exports}
\alias{\%>\%}
\title{topogRam exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the topogRam package to avoid listing the magrittr
as Depends of topogRam
}
