% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topogRam.R
\name{topogRam}
\alias{topogRam}
\title{Cartogram htmlwidget for visualizing geographical data by distorting a TopoJson topology}
\usage{
topogRam(shape, value, tooltip_label = NULL, format_value = NULL,
  palette = "Viridis", n_iteration = 20, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{shape}{An \code{sf} object.}

\item{value}{Variable name to use to distort topology.}

\item{tooltip_label}{Formula for tooltip's label.}

\item{format_value}{Character, D3 format to use, see \url{https://github.com/d3/d3-format}.}

\item{palette}{Color palette to use, see \url{https://github.com/d3/d3-scale-chromatic}, all \code{interpolate} palettes are available.}

\item{n_iteration}{Number of iterations to run the algorithm for. Higher numbers distorts the areas closer to their associated value,
at the cost of performance.}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}

\item{elementId}{Use an explicit element ID for the widget.}
}
\description{
Continuous area cartograms with `d3.js`
}
