% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topogRam.R
\name{topogRam}
\alias{topogRam}
\title{`d3.js` topogram}
\usage{
topogRam(data, key_var, shape = "france-reg", geo_id = "id",
  geo_lab = NULL, colors, origin = NULL, width = 500, height = 500,
  elementId = NULL)
}
\arguments{
\item{data}{a \code{data.frame} with at least two variables : the geo id and the value associated}

\item{key_var}{a character vector of length one or more, or a named list. The value to represent on the map}

\item{shape}{geographical shape to use, should be one of \code{france-reg}, \code{france-reg-2016}, \code{france-dep}, \code{usa-states}}

\item{geo_id}{name of variable containing the geographical id}

\item{geo_lab}{name of variable containing the geographical label}

\item{colors}{a vector of color to use on the map}
}
\description{
Continuous area cartograms with `d3.js`
}

