% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo-extras.R
\name{topogram_legend}
\alias{topogram_legend}
\title{Legend for topogram widget}
\usage{
topogram_legend(
  topo,
  colors = NULL,
  labels = NULL,
  formatter = NULL,
  title = NULL,
  direction = c("h", "v"),
  height = "250px",
  width = "250px"
)
}
\arguments{
\item{topo}{A \code{\link[=topogram]{topogram()}} / \code{\link[=topogram_proxy]{topogram_proxy()}} \code{htmlwidget} object.}

\item{colors}{Vector of colors used in legend, default is to use colors used in \code{\link[=topogram]{topogram()}}.}

\item{labels}{Labels to display for values, default is to use range of values used in \code{\link[=topogram]{topogram()}}.}

\item{formatter}{Function to format labels, like \code{\link[scales:label_number]{scales::label_number()}}.}

\item{title}{Title for the legend.}

\item{direction}{Direction: horizontal or vertical.}

\item{height, width}{Height, width for legend. For gradient legend it represent the size of the dradient according to direction.}
}
\description{
Add a gradient legend in a \code{\link[=topogram]{topogram()}} widget.
}
