% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo-extras.R
\name{topogram_legend}
\alias{topogram_legend}
\title{Legend for topogram widget}
\usage{
topogram_legend(
  topo,
  colors = NULL,
  labels = NULL,
  formatter = NULL,
  title = NULL,
  direction = c("h", "v"),
  height = "250px",
  width = "250px"
)
}
\arguments{
\item{topo}{A \code{\link[=topogram]{topogram()}} \code{htmlwidget} object.}

\item{colors}{Vector of colors.}

\item{labels}{Labels associated to colors:
\itemize{
\item For gradient: a vector of length 2 for range of values.
\item For breaks: a vector of same length as colors, or length+1 to display the minimal value.
\item For discrete: a vector of same length as colors.
}}

\item{formatter}{Function to format labels.}

\item{title}{Title for the legend.}

\item{direction}{Direction: horizontal or vertical.}

\item{height, width}{Height, width for legend. For gradient legend it represent the size of the dradient according to direction.}
}
\description{
Add a legend
}
\examples{
library(topogram)
data(paris)

topogram(
  shape = paris,
  value = "TOTAL",
  n_iteration = 10
) \%>\% add_legend(
  title = "Population",
  n_cells = 12,
  orientation = "vertical"
)


# Format labels favue with D3 format
topogram(
  shape = paris,
  value = "TOTAL",
  n_iteration = 10
) \%>\% add_legend(
  title = "Population",
  label_format = ".2s",
  orientation = "vertical"
)


# No space between cells
# and custom labels
topogram(
  shape = paris,
  value = "TOTAL",
  n_iteration = 10
) \%>\% add_legend(
  title = "Population",
  n_cells = 20,
  cells_padding = 0,
  cells_width = 10,
  cells_height = 20,
  labels = c("15 000", rep("", 18), "250 000"),
  orientation = "vertical"
)
}
