% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topogram_select.R
\name{topogram_select}
\alias{topogram_select}
\title{Select mmenu to update \code{\link[=topogram]{topogram()}}}
\usage{
topogram_select(
  topogramId,
  sfobj,
  values,
  label = "{value}",
  palette = "viridis",
  rescale_to = c(1, 1000),
  n_iteration = 10,
  width = NULL
)
}
\description{
Use in RMarkdown documents to update a \code{\link[=topogram]{topogram()}} dynamically.
}
